fix errors due to the recent changes from int to size_t i.e the colors weren't displayed anymore

+++ Code/GUI/lib/QTableWidgetChild.cxx
@@ -229,7 +229,7 @@ void QTableWidgetChild::DisplayColumnNames( QString TableName,
  std::list<std::string> ColumnNames)
 {
   size_t numberCol=ColumnNames.size();
-  this->setColumnCount( static_cast< int >( numberCol ) );
+  this->setColumnCount(static_cast<int>(numberCol));
 
   int i = 0;
   for( std::list<std::string>::iterator iter = ColumnNames.begin();
@@ -287,7 +287,7 @@ void QTableWidgetChild::DisplayContent(GoDBTableWidgetContainer* iLinkToRowConta
         {
         if (NbofRows == 0)
           {
-          NbofRows = static_cast< unsigned int >( RowContainer[i].second.size() );
+          NbofRows = RowContainer[i].second.size();
           this->setRowCount(NbofRows);
           RowsCountSet = true;
           }
